requirements-api.txt
# FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# LangChain and vector store
langchain-chroma==0.1.2
langchain-openai==0.1.8
langchain-core==0.2.0
chromadb==0.4.22

# Additional dependencies
openai==1.35.0
